Factor out cli.ts, watcher.ts.
