+++ frontend/querybuilder/querybuilder.ts
@@ -1,14 +1,10 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import * as child_process from 'child_process';
-import chokidar from 'chokidar';
-import chalk from 'chalk';
 
 import {
   strContains,
   getGraphQlFragments,
-  argvHasOption,
-  debouncer
 } from "./util";
 
 /**
@@ -20,17 +16,17 @@ import {
  * The path to the directory where we'll get our GraphQL queries, and
  * where we'll put our generated TypeScript source code.
  */
-const LIB_PATH_PARTS = ['frontend', 'lib', 'queries'];
+export const LIB_PATH_PARTS = ['frontend', 'lib', 'queries'];
 export const LIB_PATH = path.join(...LIB_PATH_PARTS);
 
 /** The path into which Apollo codegen:generate will put its generated files. */
 const GEN_PATH = path.join(LIB_PATH, '__generated__');
 
 /** The location of the GraphQL schema for the server. */
-const SCHEMA_PATH = path.join('schema.json');
+export const SCHEMA_PATH = path.join('schema.json');
 
 /** The extension for all our GraphQL queries/mutations/fragments. */
-const DOT_GRAPHQL = '.graphql';
+export const DOT_GRAPHQL = '.graphql';
 
 /**
  * Files generated by Apollo codegen:generate that we want to copy over to
@@ -219,7 +215,7 @@ export class GraphQlFile {
  * A custom error that indicates an error from this tool, which
  * users can take steps to resolve.
  */
-class ToolError extends Error {}
+export class ToolError extends Error {}
 
 /**
  * Determine whether we need to run Apollo codegen:generate, based on
@@ -295,54 +291,14 @@ export function runApolloCodegen(force: boolean = false): number {
 }
 
 /** Options for our main querybuilder functionality. */
-interface MainOptions {
+export interface MainOptions {
   /** Whether to force-run Apollo codegen:generate even if we don't think it's necessary. */
   forceApolloCodegen: boolean;
 }
 
-/** Watch GraphQL queries and schema and re-build queries when they change. */
-function watch(options: MainOptions, debounceMs = 250) {
-  const myFile = path.basename(process.argv[1]);
-  const myFileContents = fs.readFileSync(myFile);
-  const paths = [
-    SCHEMA_PATH,
-    path.posix.join(...LIB_PATH_PARTS, `*${DOT_GRAPHQL}`)
-  ];
-  const checkIfMyFileHasChanged = () => {
-    if (!myFileContents.equals(fs.readFileSync(myFile))) {
-      console.log(chalk.yellowBright(
-        `WARNING: ${myFile} has changed, consider pressing Ctrl-C\n` +
-        `and re-running this command.`
-      ));
-    }
-  };
-  chokidar.watch([myFile], {
-    ignoreInitial: true,
-    awaitWriteFinish: true
-  }).on('all', checkIfMyFileHasChanged);
-  chokidar.watch(paths).on('all', debouncer(() => {
-    let exitCode = 1;
-
-    try {
-      exitCode = main(options);
-    } catch (e) {
-      if (e instanceof ToolError) {
-        console.log(e.message);
-      } else {
-        throw e;
-      }
-    }
-
-    if (exitCode !== 0) {
-      console.log(chalk.redBright('ERROR: Rebuilding GraphQL queries failed!'));
-    }
-    console.log(`Waiting for changes in ${paths.join(', ')}...`);
-    checkIfMyFileHasChanged();
-  }, debounceMs));
-}
 
 /** Our main query-building functionality. */
-function main(options: MainOptions): number {
+export function main(options: MainOptions): number {
   const apolloStatus = runApolloCodegen(options.forceApolloCodegen);
   if (apolloStatus !== 0) {
     return apolloStatus;
@@ -361,25 +317,4 @@ function main(options: MainOptions): number {
 
   console.log('\nDone!');
   return 0;
-}
-
-if (!module.parent) {
-  if (argvHasOption('-h', '--help')) {
-    console.log(`usage: ${process.argv[1]} [OPTIONS]\n`);
-    console.log(`options:\n`);
-    console.log('  -f / --force   Force run Apollo Codgen');
-    console.log('  -w / --watch   Watch files for changes');
-    console.log('  -h / --help    Show this help');
-    process.exit(0);
-  }
-
-  const mainOptions = {
-    forceApolloCodegen: argvHasOption('-f', '--force')
-  };
-
-  if (argvHasOption('-w', '--watch')) {
-    watch(mainOptions);
-  } else {
-    process.exit(main(mainOptions));
-  }
 }
